AP-79: updated checkout form
